+++ app/library/page.tsx
@@ -1,10 +1,5 @@
-import AppShell from "@/components/layout/app-shell";
-import LibraryDashboard from "@/components/library/library-dashboard";
+import { redirect } from "next/navigation";
 
 export default function LibraryPage() {
-  return (
-    <AppShell>
-      <LibraryDashboard />
-    </AppShell>
-  );
+  redirect("/resources");
 }